# Conflicts:
#	lib/cuckoo/common/web_utils.py
#	modules/processing/curtain.py
#	web/analysis/views.py
#	web/api/views.py
#	web/submission/views.py
#	web/templates/api/index.html

+++ lib/cuckoo/common/demux.py
@@ -207,7 +207,7 @@ def demux_sflock(filename, options):
     except Exception as e:
         log.error(e, exc_info=True)
 
-    if cuckoo_conf.cuckoo.delete_archive:
+    if cuckoo_conf.cuckoo.delete_archive and retlist:
         os.remove(filename)
 
     return retlist
